Add instructions to run with different hostname
I was using this but was unable to get it working from a different machine. I found the answer was to set CORSPROXY_HOST by looking at the source code. This isn't documented anywhere but I think it is a common enough usage to require explanation. This would have saved me some time.